scripts/requirements.txt
psycopg2-binary
minio